feat(marketplace-duck): refactor marketplace duck
