+++ src/common/marketplaces/types.js
@@ -1,0 +1,27 @@
+export const marketplacesTypes = {
+	MARKETPLACE_TRANSACTIONS_LOAD: 'marketplace/transactions/LOAD',
+	MARKETPLACE_TRANSACTIONS_ADD: 'marketplace/transactions/ADD',
+	MARKETPLACE_TRANSACTIONS_SET: 'marketplace/transactions/SET',
+	MARKETPLACE_TRANSACTIONS_UPDATE_ONE: 'marketplace/transactions/UPDATE_ONE',
+	MARKETPLACE_STAKES_LOAD: 'marketplace/stakes/LOAD',
+	MARKETPLACE_STAKES_UPDATE_ONE: 'marketplace/stakes/UPDATE_ONE',
+	MARKETPLACE_STAKES_SET: 'marketplace/stakes/SET',
+	MARKETPLACE_STAKES_WITHDRAW: 'marketplace/stakes/WITHDRAW',
+	MARKETPLACE_STAKES_PLACE: 'marketplace/stakes/PLACE',
+	MARKETPLACE_TRANSACTIONS_UPDATE_STATUS: 'marketplace/transactions/UPDATE_STATUS',
+	MARKETPLACE_TRANSACTIONS_STAKE_START: 'marketplace/transactions/STAKE_START',
+	MARKETPLACE_TRANSACTIONS_STAKE_CONFIRM: 'marketplace/transactions/STAKE_CONFIRM',
+	MARKETPLACE_TRANSACTIONS_WITHDRAW_START: 'marketplace/transactions/WITHDRAW_START',
+	MARKETPLACE_TRANSACTIONS_WITHDRAW_CONFIRM: 'marketplace/transactions/WITHDRAW_CONFIRM',
+	MARKETPLACE_TRANSACTIONS_CURRENT_UPDATE: 'marketplace/transactions/current/UPDATE',
+	MARKETPLACE_TRANSACTIONS_CURRENT_SET: 'marketplace/transactions/current/SET',
+	MARKETPLACE_TRANSACTIONS_CURRENT_CANCEL: 'marketplace/transactions/current/CANCEL',
+	MARKETPLACE_TRANSACTIONS_CURRENT_CLEAR: 'marketplace/transactions/current/CLEAR',
+	MARKETPLACE_POPUP_SHOW: 'marketplace/popup/show',
+	MARKETPLACE_STATE_SHOW: 'marketplace/state/show',
+	MARKETPLACE_RP_LOAD: 'marketplace/rp/load',
+	MARKETPLACE_RP_UPDATE: 'marketplace/rp/update',
+	MARKETPLACE_RP_APPLICATION_ADD: 'marketplace/rp/application/add'
+};
+
+export default marketplacesTypes;
